looks like the dims match now. Crux was: hb = round((float)h / (float)binSize); wb = round((float)w / (float)binSize);  -- for Piotr's code.

+++ reference_code/piotr_fhog/private/gradientMex.cpp
@@ -394,7 +394,7 @@ void mGradHist( int nl, mxArray *pl[], int nr, const mxArray *pr[] ) {
     clipHog  = (nr>=7) ? (float) mxGetScalar(pr[6])    : 0.2f;
     full     = (nr>=8) ? (bool) (mxGetScalar(pr[7])>0) : false;
     //hb=h/binSize; wb=w/binSize; //Piotr's orig
-    hb = round(h/binSize); wb = round(w/binSize); // Forrest (to match VOC5?)
+    hb = round((float)h / (float)binSize); wb = round((float)w / (float)binSize); // Forrest (to match VOC5?)
     if( useHog==0 ) {
         pl[0] = mxCreateMatrix3(hb,wb,nOrients,mxSINGLE_CLASS,1,(void**)&H);
         gradHist( M, O, H, h, w, binSize, nOrients, softBin, full );
